% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUdiff.R
\name{computeUdiff}
\alias{computeUdiff}
\title{computeUdiff}
\usage{
computeUdiff(u, u0 = NULL)
}
\arguments{
\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{u0}{the given condition to be estimated. The default is NULL.
If u0 is NULL, then u0 = u.}
}
\value{
a matrix of difference of u and u0
}
\description{
computeUdiff
}
\details{
Supposing u and u0 are vectors with length n and n0 respectively,
this function will return a n0*n-size matrix, the i-th row of it is equal to
u - u0_i
}
\examples{
\dontrun{

u <- 1:5
# Default u0
computeUdiff(u)
# Given u0
u0 <- seq(1.5, 4.5, by = 1)
computeUdiff(u, u0)
}
}
\seealso{
\code{\link{kernel_weight}}
}
