% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cov_Est.R
\name{Cov_Est}
\alias{Cov_Est}
\title{Cov_Est}
\usage{
Cov_Est(Y, weight)
}
\arguments{
\item{Y}{the p*n size matrix}

\item{weight}{the weight at u0}
}
\value{
the covariance matrix
}
\description{
this routine computes the estimator of covariance
}
