% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cov_EstMu.R
\name{Cov_EstMu}
\alias{Cov_EstMu}
\title{Cov_EstMu}
\usage{
Cov_EstMu(Y, u, u_center, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the p*n size matrix}

\item{u}{the condition}

\item{u_center}{the given center}

\item{h}{the bandwidth}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the covariance matrix
}
\description{
this routine computes the estimator of covariance
}
