% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovDelete.R
\name{CovDelete}
\alias{CovDelete}
\title{CovDelete}
\usage{
CovDelete(Y_del, weight_del)
}
\arguments{
\item{Y_del}{the matrix when the i-th column is deleted}

\item{weight_del}{the vector when the i-th entry is deleted}
}
\value{
covariance estimator
}
\description{
This routine computes the covariance when the ith observation
is deleted atcenter u(i).
}
