% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovDeleteDCM.R
\name{CovDeleteDCM}
\alias{CovDeleteDCM}
\title{CovDeleteDCM}
\usage{
CovDeleteDCM(Y, Aver_i, Residual_i, Weight_i, Alpha_i)
}
\arguments{
\item{Y}{the p*n size matrix}

\item{Aver_i}{the Average of Y_i at center u(i), it comes from the results
of function Average, p * 1 matrix}

\item{Residual_i}{the residual of Y_i - Aver_i, this can be calculted
outside the loop.}

\item{Weight_i}{the 1 * n weight matrix, it represents the weight for
the u(i), it is a sparse matrix.}

\item{Alpha_i}{the coefficient of alpha(i,u(i)), see my technical reports
for details.}
}
\value{
the covariance when deleting the ith observation
}
\description{
This routine computes the covariance when the ith observation is
 deleted at center u(i). We use some algebra techniques to calculate the
 covariance estimation atcenter u(i) when the ith observation is deleted.
 This can avoid the loop of deleting ith observation, which can reduce
 the computation complexity. Actually, we can use the former computation
 results to estimate the covariance avoiding the computation of new dataset
 without ith observation. Note that, we use the covariance format in DCM.
 See details in Dynamiccovariance model.
}
