% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Th.R
\name{CV_Th}
\alias{CV_Th}
\title{CV_Th}
\usage{
CV_Th(Yi, weight, Index.obs, r)
}
\arguments{
\item{Yi}{the p * n matrix}

\item{weight}{the weight at u0}

\item{Index.obs}{the index of observation}

\item{r}{the threshold value}
}
\value{
the value of cross validation function
}
\description{
This routine computes the goal function for optimal threshold value
}
\examples{
\dontrun{

data(u)
data(Yresid_lc)
data(Index.obs)
data(hdcm1)
u <- as.vector(u)

u0_len <- 25
u0 <- seq(-0.9, 0.9, length.out = u0_len)
a <- 0
bcoef <- 0.6
thdcm1  <- rep(0, u0_len)
Weight <- kernel_weight(u, u0 = u0, ktype = 'gaussian', bw = hdcm1$minimum)
for (i in 1:u0_len) {
    b1 <- bcoef * max(Cov_Est(Y = Yresid_lc, Weight[i,]))
    thdcm1_temp <- optimise(CV_Th, c(a, b1), tol = 1e-6, Yi = Yresid_lc,
                  weight = Weight[i,], Index.obs = Index.obs)
    thdcm1[i] <-thdcm1_temp$minimum
}

# NCM, threshold value selection
data(Ystd_LC)
data(hncm)
thncm  <- rep(0, u0_len)
Weight <- kernel_weight(u, u0 = u0, ktype = 'gaussian', bw = hncm$minimum)
for (i in 1:u0_len) {
    b1 <- bcoef * max(Cov_Est(Y = Ystd_LC, Weight[i,]))
    thncm_temp <- optimise(CV_Th, c(a, b1), tol = 1e-8, Yi = Ystd_LC,
                  weight = Weight[i,], Index.obs = Index.obs)
    thncm[i] <-thncm_temp$minimum
}
}
}
\seealso{
\code{\link{DCMCVh}}, \code{\link{Cov_Est}}
}
