% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVHstdLL_LH.R
\name{CVHstdLL_LH}
\alias{CVHstdLL_LH}
\title{CVHstdLL_LH}
\usage{
CVHstdLL_LH(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{u}{the condition, row matrix, size 1 * n1.}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the value of cross validation
}
\description{
This function is used to compute the delete-i cross validation value
}
\examples{
\dontrun{

data(Yresid_ll)
data(u)
data(LowerBoundary)
h_grid <- matrix(seq(0.5, 2, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, CVHstdLL_LH, Y = Yresid_ll, u = u)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# select the optimal bandwidth for diagonal entries of covariance
hstd_LL_LH <- optimise(CVHstdLL_LH, c(LowerBoundary, 2), tol = 1e-6,
                   Y = Yresid_ll, u = u)
abline(v = hstd_LL_LH$minimum, col="red")
}
}
\seealso{
\code{\link{CVLL}}
}
