% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVHstdLL_LS.R
\name{CVHstdLL_LS}
\alias{CVHstdLL_LS}
\title{CVHstdLL_LS}
\usage{
CVHstdLL_LS(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the cross validation value
}
\description{
This function calculates the cross validation values using
local linear smoother when bandwidth is given.
}
\examples{
\dontrun{

data(Y)
data(u)
data(LowerBoundary)
# find the optimal bandwidth for mean estimation
h_mean_ll <- optimise(CVLL, c(LowerBoundary, 2), tol = 1e-6, Y = Y, u = u)
# compute the residual
YResidual_LL <- YResidualLL(Y = Y, u = u, h = h_mean_ll$minimum)
h_grid <- matrix(seq(1.5, 2, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, CVHstdLL_LS, Y = YResidual_LL, u = u)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# select the optimal bandwidth for diagonal entries of covariance
h_std_ll_ls <- optimise(CVHstdLL_LS, c(LowerBoundary, 2), tol = 1e-6,
                   Y = YResidual_LL_LS, u = u)
abline(v = h_std_ll_ls$minimum, col="red")
}
}
\seealso{
\code{\link{CVLL}}
}
