% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVLL.R
\name{CVLL}
\alias{CVLL}
\title{CVLL}
\usage{
CVLL(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the cross validation value
}
\description{
This function compute the cross validation value of local linear estimator
using the Gaussian kernel. We trimmed  5% percent of upper and lower
observations to avoid boundary effects.
}
