% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvRhoAll.R
\name{CvRhoAll}
\alias{CvRhoAll}
\title{CvRhoAll}
\usage{
CvRhoAll(yadd, ytime, u, h, ktype = "gaussian")
}
\arguments{
\item{u}{the conditon}

\item{h}{the bandwidth}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the value of cross validation
}
\description{
CvRhoAll
}
