% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMCVh.R
\name{DCMCVh}
\alias{DCMCVh}
\title{DCMCVh}
\usage{
DCMCVh(Y, u, h, Index.y)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}

\item{Index.y}{the index of subset-y-variables}
}
\value{
the value of cross validation
}
\description{
This routine calculate the cross validation value without
threshold andshrinkage. We use subset-y-variables method.
}
\examples{
\dontrun{

data(u)
data(LowerBoundary)
data(Yresid_lc)
data(Index.y)
h_grid <- matrix(seq(0.2, 1, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, DCMCVh, Y = Yresid_lc, u = u, Index.y = Index.y)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# select the optimal bandwidth for diagonal entries of covariance
hdcm1 <- optimise(DCMCVh, c(LowerBoundary, 2), tol = 1e-6,
                   Y = Yresid_lc, u = u, , Index.y = Index.y)
abline(v = hdcm1$minimum, col="red")
}
}
\seealso{
\code{\link{CVLC}}
}
