% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMCVhMu.R
\name{DCMCVhMu}
\alias{DCMCVhMu}
\title{DCMCVhMu}
\usage{
DCMCVhMu(Y, u, h, Y_res_h1, Index.y)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}

\item{Index.y}{the index of subset-y-variables}
}
\value{
the value of cross validation
}
\description{
This routine calculate the cross validation value without
threshold andshrinkage. We use subset-y-variables method.
}
\examples{
\dontrun{

data(u)
data(LowerBoundary)
data(Yresid_lc)
data(Index.y)
data(Y)
h_grid <- matrix(seq(0.1, 1, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, DCMCVhMu, Y = Y, u = u,
            Y_res_h1 = Yresid_lc, Index.y = Index.y)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# select the optimal bandwidth for diagonal entries of covariance
hdcm2 <- optimise(DCMCVhMu, c(LowerBoundary, 2), tol = 1e-6,
                   Y = Y, u = u, Y_res_h1 = Yresid_lc, Index.y = Index.y)
abline(v = hdcm2$minimum, col="red")
}
}
\seealso{
\code{\link{CVLC}}
}
