% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_bh.R
\name{fdr_bh}
\alias{fdr_bh}
\title{fdr_bh}
\usage{
fdr_bh(pvals, q = 0.05, method = "pdep")
}
\arguments{
\item{pvals}{A vector  containing the p-value of each individual test
in a family of tests.}

\item{q}{The desired false discovery rate. The default is 0.05.}

\item{method}{'pdep' or 'dep'. If 'pdep,' the original Bejnamini & Hochberg
FDR procedure is used, which is guaranteed to be accurate if
the individual tests are independent or positively dependent
(e.g., Gaussian variables that are positively correlated or
independent).  If 'dep,' the FDR procedure described in
Benjamini & Yekutieli (2001) that is guaranteed to be accurate
for any test dependency structure (e.g., Gaussian variables
with any covariance matrix) is used. 'dep' is always
appropriate to use but is less powerful than 'pdep.'
The default is 'pdep'.}
}
\value{
A binary vector of the same size as the input "pvals." If the
        ith element of h is 1, then the test that produced the ith p-value
        in pvals is significant
        (i.e., the null hypothesisof the test is rejected).
}
\description{
Executes the \insertCite{benjaminiControllingFalseDiscovery1995;textual}{llfdr}
          and the \insertCite{benjaminiControlFalseDiscovery2001;textual}{llfdr} procedure for controlling the false discovery
            rate (FDR) of a family of hypothesis tests. FDR is the expected
            proportion of rejected hypotheses that are mistakenly rejected
            (i.e., the null hypothesis is actually true for those tests).
            FDR is a somewhat less conservative/more powerful method for
            correcting for multiple comparisons than procedures like Bonferroni
            correction that provide strong control of the family-wise
            error rate (i.e., the probability that one or more null
             hypotheses are mistakenly rejected).
}
\references{
\insertRef{benjaminiControllingFalseDiscovery1995}{llfdr}

\insertRef{benjaminiControlFalseDiscovery2001}{llfdr}
}
