% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLR.R
\name{GLR}
\alias{GLR}
\title{GLR}
\usage{
GLR(res_linear, res_nonlinear, h, len_interval, n)
}
\arguments{
\item{res_linear}{the residual of linear model}

\item{res_nonlinear}{the residual of nonlinear model}

\item{h}{the bandwidth}

\item{len_interval}{the range of u}

\item{n}{sample size}
}
\value{
the p value of chi square test
}
\description{
this routine computes the p value using generalized likelihood
ratio statistics \insertCite{Fan2001a}{llfdr}
}
\references{
\insertRef{Fan2001a}{llfdr}
}
