% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h1_ll}
\alias{h1_ll}
\title{Bandwidth}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(h1_ll)
}
\description{
The bandwidth was computed by \code{\link{CVLL}}
}
\keyword{datasets}
