% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdcm1}
\alias{hdcm1}
\title{The bandwidth}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(hdcm1)
}
\description{
The bandwidth was computed by \code{\link{DCMCVh}}
}
\keyword{datasets}
