% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdcm2}
\alias{hdcm2}
\title{The bandwidth}
\format{An object of class \code{list} of length 2.}
\usage{
data(hdcm2)
}
\description{
The bandwidth was computed by \code{\link{DCMCVhMu}}
}
\keyword{datasets}
