% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hncm}
\alias{hncm}
\title{The bandwidth}
\format{An object of class \code{list} of length 2.}
\usage{
data(hncm)
}
\description{
The bandwidth was computed by \code{\link{NCMCVh}}
}
\keyword{datasets}
