% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hstd_lc}
\alias{hstd_lc}
\title{The bandwidth}
\format{An object of class \code{list} of length 2.}
\usage{
data(hstd_lc)
}
\description{
The bandwidth was computed by \code{\link{CVHstdLC}}
}
\keyword{datasets}
