% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hstd_LL_LH}
\alias{hstd_LL_LH}
\title{The bandwidth}
\format{An object of class \code{list} of length 2.}
\usage{
data(hstd_LL_LH)
}
\description{
The bandwidth was computed by \code{\link{CVHstdLL_LH}}
}
\keyword{datasets}
