% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hstd_LL_LS}
\alias{hstd_LL_LS}
\title{The bandwidth}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(hstd_LL_LS)
}
\description{
The bandwidth was computed by \code{\link{CVHstdLL_LS}}
}
\keyword{datasets}
