% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Index.obs}
\alias{Index.obs}
\title{Index of observations}
\format{An object of class \code{matrix} with 78 rows and 100 columns.}
\usage{
data(Index.obs)
}
\description{
The subset of observations, size: 78 * 100, is used for thresholding.
 For more details, see Bickel, P. J. and Levina, E. (2008).
 Covariance regularization by thresholding. The Annals of Statistics, 36(6):2577–2604.
}
\keyword{datasets}
