% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Index.y}
\alias{Index.y}
\title{Index of subset-y}
\format{An object of class \code{matrix} with 4 rows and 50 columns.}
\usage{
data(Index.y)
}
\description{
The subset of y variables, size: 4 * 50. For more details, see
 Chen, Z. and Leng, C., 2016. Dynamic covariance models.
 Journal of the American Statistical Association, 111(515), pp.1196-1207.
}
\keyword{datasets}
