% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{interpolate}
\usage{
interpolate(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector which each entry is non-negative
}
\description{
this function inserts the positive value into the entries where
the value is negative. If all of the vector are negative, then replaces each
entry with 1e-4.
}
