% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCompute.R
\name{kernelCompute}
\alias{kernelCompute}
\title{kernelCompute}
\usage{
kernelCompute(u_diff, ktype = "gaussian", bw = 0.1)
}
\arguments{
\item{u_diff}{a matrix of difference of u and u0}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}

\item{bw}{bandwidth, default is 0.1}
}
\value{
kernel matrix
}
\description{
kernelCompute
}
\examples{
\dontrun{

n <- 20
n0 <- 10
u <- seq(-1, 1, length.out = n)
u0 <- seq(-0.9, 0.9, length.out = n0)
u_diff <- computeUdiff(u, u0)
# Gaussian kernel with bandwidth 0.1, default
kernel <- kernelCompute(u_diff)
# Cosine kernel with bandwidth 0.4
kernel_cos <- kernelCompute(u_diff, ktype = 'cosine', bw = 0.4)
}
}
\seealso{
\code{\link{computeUdiff}}, \code{\link{kernel_weight}}
}
