% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEBeta2.R
\name{LMEBeta2}
\alias{LMEBeta2}
\title{LMEBeta2}
\usage{
LMEBeta2(v, y2, expu, k, uk, SK, SUK)
}
\arguments{
\item{v}{the beta when the i-th observation deleted}

\item{y2}{the square of y}

\item{expu}{the exp(u_j - u_i)}

\item{k}{the kernel}

\item{uk}{the (u_j - u_i) * kernel}

\item{SK}{summation of kernel}

\item{SUK}{summation of kernel * u}
}
\value{
goal function value
}
\description{
This function is used to compute the delete-i beta goal
function f = (sbuk / sbk - suk / sk)^2;
}
