% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEDiagEstUFmin.R
\name{LMEDiagEstUFmin}
\alias{LMEDiagEstUFmin}
\title{LMEDiagEstUFmin}
\usage{
LMEDiagEstUFmin(Y, u, u0, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{u0}{the given condition to be estimated. The default is NULL.
If u0 is NULL, then u0 = u.}

\item{h}{the bandwidth, scalar}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the estimator of  diagonal entries at each given u0.
}
\description{
this routine computes the corresponding sigma matrix
with standard bandwidth at each u_i using local linear smoother.
}
\examples{
\dontrun{

data(Yresid_ll)
data(u)
data(hstd_LL_LH)
# compute the estimator of diagonal entries
Sigma_LL_LH <- SigmaHstdLL_LH(Y = Yresid_ll, u = u, h = hstd_LL_LH$minimum)
Ystd_LL_LH = Yresid_ll / sqrt(Sigma_LL_LH)
}
}
\seealso{
\code{\link{CVLL}},\code{\link{CVHstdLL}}
}
