% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCMCVh.R
\name{NCMCVh}
\alias{NCMCVh}
\title{NCMCVh}
\usage{
NCMCVh(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation, p * n matrix}

\item{u}{the condtion, 1 * n matrix}

\item{h}{the bandwidth, scalar}
}
\value{
the value of cross validation function
}
\description{
This routine calculate the cross validation value using NCM method.
}
