% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimise_LMEBeta2.R
\name{Optimise_LMEBeta2}
\alias{Optimise_LMEBeta2}
\title{Optimise_LMEBeta2}
\usage{
Optimise_LMEBeta2(y2, expu, mat_uk, r0)
}
\arguments{
\item{y2}{the square of y}

\item{expu}{the exp(u_j - u_i)}

\item{mat_uk}{the matrix, the first row is k, the secound row is u * k}

\item{r0}{the ratio of SUK to SK}
}
\value{
the optimal bandwidth
}
\description{
optimise function LMEBeta2
}
