% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhoEstAll.R
\name{RhoEstAll}
\alias{RhoEstAll}
\title{RhoEstAll}
\usage{
RhoEstAll(yadd, ytime, u, u_test, h, ktype = "gaussian")
}
\arguments{
\item{ktype}{}
}
\value{
the correlation coefficient matrix
}
\description{

}
