% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrinkage.R
\name{Shrinkage}
\alias{Shrinkage}
\title{Shrinkage}
\usage{
Shrinkage(Y, w, Cov, Ind_t0)
}
\arguments{
\item{Y}{the data, p * n matrix}

\item{w}{the weight}

\item{Cov}{the covariance when the i-th deleted}

\item{Ind_t0}{the indicator matrix}
}
\value{
the shrinkage covariance
}
\description{
This function computes the shrinkage covariance after threshold .
}
