% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_Estu.R
\name{Sigma_Estu}
\alias{Sigma_Estu}
\title{Sigma_Estu}
\usage{
Sigma_Estu(Y, u, h, u_test, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}

\item{u_test}{the condition u for testing}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the diagonal entries estimator
}
\description{
Sigma_Estu
}
