% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_Gen_Model1.R
\name{Sigma_Gen_Model1}
\alias{Sigma_Gen_Model1}
\title{Sigma_Gen_Model1}
\usage{
Sigma_Gen_Model1(p, u)
}
\arguments{
\item{p}{the number of y variables}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}
}
\value{
p * p * length(u) size matrix
}
\description{
This runtine generates the p * p * length(u) size Sigma matrix.
}
