% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_hstd_lc}
\alias{Sigma_hstd_lc}
\title{The Estimator of Diagonal Entries}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 100 columns.
}
\usage{
data(Sigma_hstd_lc)
}
\description{
The estimator of diagonal entries was computed by \code{\link{SigmaHstdLC}}
}
\keyword{datasets}
