% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_LL_LS}
\alias{Sigma_LL_LS}
\title{The Estimator of Diagonal Entries}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 100 columns.
}
\usage{
data(Sigma_LL_LS)
}
\description{
The estimator of diagonal entries was computed by \code{\link{SigmaHstdLL_LS}}
}
\keyword{datasets}
