% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaHstdLC.R
\name{SigmaHstdLC}
\alias{SigmaHstdLC}
\title{SigmaHstdLC}
\usage{
SigmaHstdLC(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the estimator of  diagonal entries at each given u_i.
}
\description{
this routine computes the corresponding sigma matrix
with standard bandwidth at each u_i.
}
\examples{
\dontrun{

data(YResidual_LC)
data(u)
data(h_std_lc)
# compute the estimator of diagonal entries
Sigma_hstd_lc <- SigmaHstdLC(Y = YResidual_LC, u = u, h = h_std_lc$minimum)
}
}
\seealso{
\code{\link{CVLC}},\code{\link{CVHstdLC}}
}
