% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaHstdLL_LS.R
\name{SigmaHstdLL_LS}
\alias{SigmaHstdLL_LS}
\title{SigmaHstdLL_LS}
\usage{
SigmaHstdLL_LS(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the estimator of  diagonal entries at each given u_i.
}
\description{
this routine computes the corresponding sigma matrix
with standard bandwidth at each u_i using local linear smoother.
}
\examples{
\dontrun{

data(Yresid_ll)
data(u)
data(hstd_LL_LS)
# compute the estimator of diagonal entries
Sigma_LL_LS <- SigmaHstdLL_LS(Y = Yresid_ll, u = u, h = hstd_LL_LS$minimum)
Ystd_LL_LS = Yresid_ll / sqrt(Sigma_LL_LS)
}
}
\seealso{
\code{\link{CVLL}},\code{\link{CVHstdLL}}
}
