% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thdcm1}
\alias{thdcm1}
\title{The threshold value}
\format{An object of class \code{numeric} of length 25.}
\usage{
data(thdcm1)
}
\description{
The threshold value was computed by \code{\link{CV_Th}}
}
\keyword{datasets}
