% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u}
\alias{u}
\title{The condition}
\format{An object of class \code{matrix} with 1 rows and 100 columns.}
\usage{
data(u)
}
\description{
A condition vector with 100 length.
}
\keyword{datasets}
