% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{The data}
\format{A data \code{matrix} with 50 rows and 100 columns.}
\usage{
data(Y)
}
\description{
This R package is translated from our initial MATLAB code.
In order to keep functions performimg consistently. We load only one small dataset saved by
MATLAB. Here, Y is a data matrix with 50 rows and 100 columns.
}
\keyword{datasets}
