% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yresid_lc}
\alias{Yresid_lc}
\title{The residual}
\format{
An object of class \code{matrix} with 50 rows and 100 columns.
}
\usage{
data(Yresid_lc)
}
\description{
The residual was computed by \code{\link{YResidualH1LC}}
}
\keyword{datasets}
