% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Yresid_ll}
\alias{Yresid_ll}
\title{The residual}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 100 columns.
}
\usage{
data(Yresid_ll)
}
\description{
The residual was computed by \code{\link{YResidualH1LL}}
}
\keyword{datasets}
