% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YResidualH1LL.R
\name{YResidualH1LL}
\alias{YResidualH1LL}
\title{YResidualH1LL}
\usage{
YResidualH1LL(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the residuals
}
\description{
This function firstly estimates the local linear weighted
average at each center u(i) given bandwidth h. Here we use gaussian kerenel.
The return is the residual of Yi - \hat{Y}_{i}.
}
\examples{
\dontrun{

data(Y)
data(u)
data(h1_ll)
Yresid_ll <- YResidualH1LL(Y = Y, u = u, h = h1_ll$minimum)
}
}
\seealso{
\code{\link{YResidualH1LC}}
}
