% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YResidualLC.R
\name{YResidualLC}
\alias{YResidualLC}
\title{YResidualLC}
\usage{
YResidualLC(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the residuals
}
\description{
This function firstly estimates the local constant weighted
average at each center u(i) given bandwidth h. Here we use gaussian kerenel.
The return is the residual of Yi - \hat{Y}_{i}.
}
\examples{
\dontrun{

data(Y)
data(u)
data(LowerBoundary)
# h is fixed
h <- 0.1
Residual_LC <- YResidualLC(Y = Y, u = u, h = h)
}
}
\seealso{
\code{\link{YResidualLL}}
}
