% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ystd_LL_LH}
\alias{Ystd_LL_LH}
\title{The standardization of Y}
\format{
An object of class \code{matrix} with 50 rows and 100 columns.
}
\usage{
data(Ystd_LL_LH)
}
\description{
The standardization of Y
}
\keyword{datasets}
