% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cov_EstMu.R
\name{Cov_EstMu}
\alias{Cov_EstMu}
\title{Cov_EstMu}
\usage{
Cov_EstMu(Y, weight)
}
\arguments{
\item{Y}{the p*n size matrix}

\item{weight}{the bandwidth at u0}
}
\value{
the covariance matrix
}
\description{
this routine computes the estimator of covariance
}
