% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Th.R
\name{CV_Th}
\alias{CV_Th}
\title{CV_Th}
\usage{
CV_Th(Yi, u, Index.obs, u_test, h, r, ktype = "gaussian")
}
\arguments{
\item{Yi}{the p * n matrix}

\item{u}{the condition u}

\item{Index.obs}{the index of observation}

\item{u_test}{the given u}

\item{h}{the optimal bandwidth}

\item{r}{the threshold value}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the value of cross validation function
}
\description{
This routine computes the goal function for optimal threshold value
}
