% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Th.R
\name{CV_Th}
\alias{CV_Th}
\title{CV_Th}
\usage{
CV_Th(Yi, u, Index.obs, u0, h, r, ktype = "gaussian")
}
\arguments{
\item{Yi}{the p * n matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{Index.obs}{the index of observation}

\item{u0}{the given condition to be estimated. The default is NULL.
If u0 is NULL, then u0 = u.}

\item{h}{the  bandwidth}

\item{r}{the threshold value}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the value of cross validation function
}
\description{
This routine computes the goal function for optimal threshold value
}
