% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVHstdLC.R
\name{CVHstdLC}
\alias{CVHstdLC}
\title{CVHstdLC}
\usage{
CVHstdLC(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{h}{the bandwidth, scalar}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the cross validation value
}
\description{
This function calculates the cross validation values using local constant
method when bandwidth is given.
}
\examples{
\dontrun{

data(u)
data(LowerBoundary)
data(Yresid_lc)
h_grid <- matrix(seq(0.2, 1, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, CVHstdLC, Y = Yresid_lc, u = u)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# select the optimal bandwidth for diagonal entries of covariance
hstd_lc <- optimise(CVHstdLC, c(LowerBoundary, 2), tol = 1e-6,
                   Y = Yresid_lc, u = u)
abline(v = hstd_lc$minimum, col="red")
}
}
\seealso{
\code{\link{CVLC}}
}
