% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVHstdLC.R
\name{CVHstdLC}
\alias{CVHstdLC}
\title{CVHstdLC}
\usage{
CVHstdLC(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the cross validation value
}
\description{
This function calculates the cross validation values using local constant
method when bandwidth is given.
}
