% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVHstdLL_LS.R
\name{CVHstdLL_LS}
\alias{CVHstdLL_LS}
\title{CVHstdLL_LS}
\usage{
CVHstdLL_LS(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{h}{the bandwidth, scalar}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the cross validation value
}
\description{
This function calculates the cross validation values using
local linear smoother when bandwidth is given.
}
\examples{
\dontrun{

data(Yresid_ll)
data(u)
data(LowerBoundary)
h_grid <- matrix(seq(1.5, 2, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, CVHstdLL_LS, Y = Yresid_ll, u = u)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# select the optimal bandwidth for diagonal entries of covariance
hstd_LL_LS <- optimise(CVHstdLL_LS, c(LowerBoundary, 2), tol = 1e-6,
                   Y = Yresid_ll, u = u)
abline(v = hstd_LL_LS$minimum, col="red")
}
}
\seealso{
\code{\link{CVLL}}
}
