% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVLC.R
\name{CVLC}
\alias{CVLC}
\title{CVLC}
\usage{
CVLC(Y, u, h)
}
\arguments{
\item{Y}{the observation, p * n matrix}

\item{u}{the condtion, 1 * n matrix}

\item{h}{the bandwidth, scalar}
}
\value{
the cross validation value
}
\description{
This function compute the cross validation value of local constant estimator
using the Gaussian kernel. We trimmed  5% percent of upper and lower
observations to avoid boundary effects.
}
\examples{
\dontrun{

data(Y)
data(u)
data(LowerBoundary)
# h is fixed
h_grid <- matrix(seq(c(LowerBoundary), 0.3, length.out = 100), nrow = 100)
cv <- apply(h_grid, 1, CVLC, Y = Y, u = u)
plot(h_grid,cv, type = 'l', xlab = "Bandwidth", ylab = "CV Values", col = "blue")
# find the optimal bandwidth
h1_lc <- optimise(CVLC, c(LowerBoundary, 2), tol = 1e-6, Y = Y, u = u)
abline(v = h1_lc$minimum, col="red")
}
}
\seealso{
\code{\link{CVLL}}
}
