% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMCV_ThMu.R
\name{DCMCV_ThMu}
\alias{DCMCV_ThMu}
\title{DCMCV_ThMu}
\usage{
DCMCV_ThMu(Yi, weight, Index.obs, r)
}
\arguments{
\item{Yi}{the p * n matrix}

\item{weight}{the weight at u0}

\item{Index.obs}{the index of observation}

\item{r}{the threshold value}
}
\value{
the value of cross validation function
}
\description{
This routine computes the goal function for optimal threshold value
}
\examples{
\dontrun{

data(u)
data(Y)
data(Index.obs)
data(hdcm2)
u <- as.vector(u)

u0_len <- 25
u0 <- seq(-0.9, 0.9, length.out = u0_len)
a <- 0
bcoef <- 0.6
thdcm2  <- rep(0, u0_len)
Weight <- kernel_weight(u, u0 = u0, ktype = 'gaussian', bw = hdcm2$minimum)
for (i in 1:u0_len) {
    b2 <- bcoef * max(Cov_EstMu(Y = Y, weight = Weight[i,]))
    thdcm2_temp <- optimise(DCMCV_ThMu, c(a, b2), tol = 1e-6, Yi = Y,
                  weight = Weight[i,], Index.obs = Index.obs)
    thdcm2[i] <-thdcm2_temp$minimum
}
}
}
\seealso{
\code{\link{DCMCVhMu}}, \code{\link{Cov_EstMu}}
}
