% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_weight.R
\name{kernel_weight}
\alias{kernel_weight}
\title{Generate the kernel weight matrix}
\usage{
kernel_weight(u, u0 = u, ktype = "gaussian", bw = 0.1, poly_order = 0)
}
\arguments{
\item{u}{the condition, row matrix, size 1 * n1.}

\item{u0}{the center condition, row matrix, size 1 * n2, the default is u.}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}

\item{bw}{bandwidth, default is 0.1}

\item{poly_order}{0 or 1, 0 represents the local constant smoother
while 1 represents the local linear smoother, default is 0.}
}
\value{
the weight matrix
}
\description{
This function generated the kernel weight matrix depending on
differe parameter settings.
}
