% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEBeta2.R
\name{LMEBeta2}
\alias{LMEBeta2}
\title{LMEBeta2}
\usage{
LMEBeta2(v, y2, expu, mat_uk, r0)
}
\arguments{
\item{v}{the beta when the i-th observation deleted}

\item{y2}{the square of y}

\item{expu}{the exp(u_j - u_i)}

\item{mat_uk}{the matrix, the first row is k, the secound row is u * k}

\item{r0}{the ratio of SUK to SK}
}
\value{
goal function value
}
\description{
This function is used to compute the delete-i beta goal
function f = sbuk - sbk * r0
}
