% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LowerBoundary}
\alias{LowerBoundary}
\title{Lower Boundary}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1 rows and 1 columns.
}
\usage{
data(LowerBoundary)
}
\description{
The lower boundary of bandwidth
}
\keyword{datasets}
