% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCMCVh.R
\name{NCMCVh}
\alias{NCMCVh}
\title{NCMCVh}
\usage{
NCMCVh(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation, p * n matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{h}{the bandwidth, scalar}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the value of cross validation function
}
\description{
This routine calculate the cross validation value using NCM method.
}
