% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimise_LMEBeta2.R
\name{Optimise_LMEBeta2}
\alias{Optimise_LMEBeta2}
\title{Optimise_LMEBeta2}
\usage{
Optimise_LMEBeta2(y2, expu, k, uk, SK, SUK)
}
\arguments{
\item{y2}{the square of y}

\item{expu}{the exp(u_j - u_i)}

\item{k}{the kernel}

\item{uk}{the (u_j - u_i) * kernel}

\item{SK}{summation of kernel}

\item{SUK}{summation of kernel * u}
}
\value{
the optimal bandwidth
}
\description{
optimise function LMEBeta2
}
