% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecoverSigma_Est.R
\name{RecoverSigma_Est}
\alias{RecoverSigma_Est}
\title{RecoverSigma_Est}
\usage{
RecoverSigma_Est(Y, u, h, u0, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation, p * n matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{h}{the bandwidth, scalar}

\item{u0}{the given condition to be estimated. The default is NULL.
If u0 is NULL, then u0 = u.}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the  sigma
}
\description{
This routine compute the i-th rows corresponding sigma for
revcoverring itself.
}
