% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhoEstAll.R
\name{RhoEstAll}
\alias{RhoEstAll}
\title{RhoEstAll}
\usage{
RhoEstAll(yadd, ytime, u, u0, h, ktype = "gaussian")
}
\arguments{
\item{yadd}{the sum}

\item{ytime}{the product}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{u0}{the given condition to be estimated. The default is NULL.
If u0 is NULL, then u0 = u.}

\item{h}{the bandwidth}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the correlation coefficient matrix
}
\description{
This function computes the estimator of coefficient.
}
