% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma}
\alias{Sigma}
\title{The Covariance Matrix}
\format{
An object of class \code{matrix} with 50 rows and 50 columns.
}
\usage{
data(Sigma)
}
\description{
The underlying covariance matrix.
}
\keyword{datasets}
