% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_Estu.R
\name{Sigma_Estu}
\alias{Sigma_Estu}
\title{Sigma_Estu}
\usage{
Sigma_Estu(Y, u, h, u0, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{h}{the bandwidth, scalar}

\item{u0}{the given condition to be estimated. The default is NULL.
If u0 is NULL, then u0 = u.}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the diagonal entries estimator
}
\description{
Sigma_Estu
}
