% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaHstdLL_LH.R
\name{SigmaHstdLL_LH}
\alias{SigmaHstdLL_LH}
\title{SigmaHstdLL_LH}
\usage{
SigmaHstdLL_LH(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the estimator of  diagonal entries at each given u_i.
}
\description{
this routine computes the corresponding sigma matrix
with standard bandwidth at each u_i using local linear smoother.
}
\examples{
\dontrun{

data(Yresid_ll)
data(u)
data(hstd_LL_LH)
# compute the estimator of diagonal entries
Sigma_LL_LH <- SigmaHstdLL_LH(Y = Yresid_ll, u = u, h = hstd_LL_LH$minimum)
Ystd_LL_LH = Yresid_ll / sqrt(Sigma_LL_LH)
}
}
\seealso{
\code{\link{CVLL}},\code{\link{CVHstdLL}}
}
