% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaHstdLL_LS.R
\name{SigmaHstdLL_LS}
\alias{SigmaHstdLL_LS}
\title{SigmaHstdLL_LS}
\usage{
SigmaHstdLL_LS(Y, u, h)
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condtion}

\item{h}{the bandwidth, scalar}
}
\value{
the estimator of  diagonal entries at each given u_i.
}
\description{
this routine computes the corresponding sigma matrix
with standard bandwidth at each u_i using local linear smoother.
}
\examples{
\dontrun{

data(YResidual_LL)
data(u)
data(h_std_lc)
# compute the estimator of diagonal entries
Sigma_hstd_ll_ls <- SigmaHstdLL_LS(Y = YResidual_LL, u = u, h = h_std_ll$minimum)
}
}
\seealso{
\code{\link{CVLL}},\code{\link{CVHstdLL}}
}
