% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YResidualH1LC.R
\name{YResidualH1LC}
\alias{YResidualH1LC}
\title{YResidualH1LC}
\usage{
YResidualH1LC(Y, u, h, ktype = "gaussian")
}
\arguments{
\item{Y}{the observation matrix}

\item{u}{the condition, it is a vector, one-dimensional array or
one-dimensional row(column) matrix}

\item{h}{the bandwidth, scalar}

\item{ktype}{the kernel type, can be "gaussian", "epanech", "triweight",
"biweight", "tricube", "triangular" and "cosine",
the default of ktype is "gaussian".}
}
\value{
the residuals
}
\description{
This function firstly estimates the local constant weighted
average at each center u(i) given bandwidth h. Here we use gaussian kerenel.
The return is the residual of Yi - \hat{Y}_{i}.
}
\examples{
\dontrun{

data(Y)
data(u)
data(h1_lc)
Yresid_lc <- YResidualH1LC(Y = Y, u = u, h = h1_lc$minimum)
}
}
\seealso{
\code{\link{YResidualH1LL}}
}
